# tests/snapshots/openapi_routers.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "guguwebui routers snapshot",
    "version": "test"
  },
  "paths": {
...
    "/api/monitor/history": {
      "get": {
        "summary": "Api Monitor History",
        "description": "获取指定指标的时间序列（含降采样）。\n\nmetric: cpu / memory / network / tps / mspt / load / disk\nrange: 10m / 30m / 1h / 6h / 12h / 1d / 3d / 7d",
        "operationId": "api_monitor_history___api_monitor_history__GET",
        "parameters": [
          {
            "name": "metric",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "cpu",
                "memory",
                "network",
                "tps",
                "mspt",
                "load",
                "disk"
              ],
              "type": "string",
              "default": "cpu",
              "title": "Metric"
            }
          },
          {
            "name": "range",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "10m",
                "30m",
                "1h",
                "6h",
                "12h",
                "1d",
                "3d",
                "7d"
              ],
              "type": "string",
              "default": "1h",
              "title": "Range"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/monitor/table": {
      "get": {
        "summary": "Api Monitor Table",
        "description": "统计表数据：各指标在当前时间范围内的 avg/min/max",
        "operationId": "api_monitor_table___api_monitor_table__GET",
        "parameters": [
          {
            "name": "range",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "10m",
                "30m",
                "1h",
                "6h",
                "12h",
                "1d",
                "3d",
                "7d"
              ],
              "type": "string",
              "default": "1h",
              "title": "Range"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/pim/tasks/{task_id}": {
      "get": {
        "summary": "Api Pim Task Status",
        "description": "获取 PIM 任务状态（原 GET /pim/task_status?task_id=）",
        "operationId": "api_pim_task_status___api_pim_tasks_{task_id}__GET",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/pip/tasks": {
      "post": {
        "summary": "Api Pip Start Task",
        "description": "发起 pip 安装/卸载后台任务（原 POST /pip/install、/pip/uninstall，合并为任务创建）",
        "operationId": "api_pip_start_task___api_pip_tasks__POST",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PipTaskCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/pip/tasks/{task_id}": {
      "get": {
        "summary": "Api Pip Task Status",
        "description": "获取 pip 任务状态（原 GET /pip/task_status?task_id=，已迁移为子资源路径）\n\n任务体不再把内部 status（success/error）展开覆盖外壳键；\n状态统一为 running|completed|failed，与 PIM 任务结构一致。",
        "operationId": "api_pip_task_status___api_pip_tasks_{task_id}__GET",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiSuccessEnvelope": {
        "properties": {
          "status": {
            "type": "string",
            "const": "success",
            "title": "Status",
            "default": "success"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "data": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "type": "object",
        "title": "ApiSuccessEnvelope",
        "description": "统一成功外壳（2.3）：status/message?/data?。"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PipTaskCreateRequest": {
        "properties": {
          "action": {
            "type": "string",
            "title": "Action"
          },
          "package": {
            "type": "string",
            "title": "Package"
          }
        },
        "type": "object",
        "required": [
          "action",
          "package"
        ],
        "title": "PipTaskCreateRequest",
        "description": "发起 pip 安装/卸载任务（install/uninstall 合并为任务创建资源）"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}